requirements.txt
psycopg2
pyalchemy
panel
watchfiles
win32print
pyserial-asyncio